% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSeedCode2.R
\name{chi.stat}
\alias{chi.stat}
\title{chi.stat function}
\usage{
chi.stat(ftab)
}
\arguments{
\item{ftab}{(argument here)}
}
\value{
chi.val (return value)
}
\description{
This function... (description here)
}
\details{
This function adds 50 to the user input.

(details here; DIU)
}
\note{
(notes here; DIU)
}
\examples{
chi.stat(ftab=rbind(c(20,10,20),c(15,15,20)))
}
\references{
(references here; DIU)
}
