% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSeedCode2.R
\docType{data}
\name{fn}
\alias{fn}
\title{experimental data
This assigns the data to variables so that they can be called and read}
\format{An object of class \code{character} of length 1.}
\usage{
fn
}
\description{
experimental data
This assigns the data to variables so that they can be called and read
}
\keyword{datasets}
