% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSeedCode2.R
\name{freq.less}
\alias{freq.less}
\title{freq.less function}
\usage{
freq.less(fx, fy)
}
\arguments{
\item{fx, fy}{This is where any argument goes}
}
\value{
c(less.count, more.count) (return value here)
}
\description{
This function finds the count the x-sample observes is greater than the one from y-sample based on ranks. (description)
}
\details{
This is where any details go (delete if unnecessary)
}
\note{
This is where any additional notes go (delete if unnecessary)
}
\examples{
freq.less(fx=c(1,2,4,9,0,0,NA),fy=c(1,4,9,NA))
}
\references{
This is where any references go (delete if unnecessary)
}
