% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSeedCode2.R
\name{gen.decision}
\alias{gen.decision}
\title{gen.decision function}
\usage{
gen.decision(est.prob, effn.subsam1 = rep(5, 8), effn.subsam2 = rep(5,
  8), fn.rep = 10^3, alpha = 0.05)
}
\arguments{
\item{est.prob, effn.subsam1, effn.subsam2, fn.rep, alpha}{(argument here)}
}
\value{
result (return value)
}
\description{
This function will test to see if two functions are the same or not the same (description)
}
\details{
(details here; DIU)
}
\note{
(notes here; DIU)
}
\examples{
gen.decision(rbind(c(20,5,10,15,20,5),c(15,10,15,10,20,5)),effn.subsam1=rep(5,4),effn.subsam2=rep(5,4),fn.rep=10^3,alpha=.05)
}
\references{
(references here; DIU)
}
