% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSeedCode2.R
\name{pow.ana.gen.decision}
\alias{pow.ana.gen.decision}
\title{pow.ana.gen.decision function}
\usage{
pow.ana.gen.decision(mean.prob1 = c(0.8, 0.2, 0.5), mean.prob2 = c(0.8,
  0.2, 0.5), effn.subsam1 = c(5, 5, 5, 5), effn.subsam2 = c(5, 5, 5,
  5), N.rep = 10^3, boot.rep = 10^3)
}
\arguments{
\item{mean.prob1, mean.prob2, effn.subsam1, effn.subsam2, N.rep, boot.rep}{(arguments here)}
}
\value{
c(mean(out[,1]),sd(out[,1]),sum(out[,3]<.05)/N.rep) (return value here)
}
\description{
This function determines how powerful the test is (description)
}
\details{
(details here; DIU)
}
\note{
(notes here; DIU)
}
\examples{
fn.rep<-10^3
num.fold<-1
pow.ana.gen.decision(mean.prob1=c(.8,.2,.5),mean.prob2=c(.4,.3,.2),N.rep=10^2, boot.rep=10^3)
}
\references{
(references here; DIU)
}
