% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSeedCode2.R
\name{simu.ustat.pattern}
\alias{simu.ustat.pattern}
\title{simu.ustat.pattern function}
\usage{
simu.ustat.pattern(mean.prob.vec, effn.subs = rep(5,
  length(mean.prob.vec)), n.rep = 10^2)
}
\arguments{
\item{mean.prob.vec, effn.subs, n.rep}{(argument here)}
}
\value{
simu.tab (return value)
}
\description{
This function will simulate the use of the pattern (description here)
}
\details{
(details here; DIU)
}
\note{
(notes here; DIU)
}
\examples{
simu.ustat.pattern(mean.prob=c(.82,.1,.6))
}
\references{
(references here; DIU)
}
